+++ machine_limits.h
@@ -48,7 +48,7 @@ bool limits_homing_required (void);
 void limits_set_homing_axes (void);
 void limits_set_machine_positions (axes_signals_t cycle, bool add_pulloff);
 void limits_set_work_envelope (void);
-
+coord_data_t *limits_homing_pulloff (coord_data_t *distance);
 void limit_interrupt_handler (limit_signals_t state);
 
 axes_signals_t limit_signals_merge (limit_signals_t signals);